requirements.txt
black==23.1.0
certifi==2022.12.7
flake8==6.0.0
isort==5.12.0
pandas==1.5.3
pip-chill==1.0.1
psycopg2-binary==2.9.5
pytest==7.2.1
python-dotenv==1.0.0
